added auth

+++ src/pages/Finished.js
@@ -1,25 +1,38 @@
-import { useEffect } from "react";
+import { useEffect, useState } from "react";
 import { BsStar } from "react-icons/bs"
-import { BiSortAlt2 } from "react-icons/bi"
+// import { BiSortAlt2 } from "react-icons/bi"
 import Task from "../components/Task";
 // import AddStarredTask from "../components/AddTask";
 import InfiniteScroll from "react-infinite-scroll-component";
-import { HiOutlineSparkles } from "react-icons/hi"
+import { HiOutlineSparkles } from "react-icons/hi";
+import { useAuthContext } from "../hooks/useAuthContext";
 
 const Finished = ({ tasks, setTasks, handleCheckTrue, handleCheckFalse, handleStarredTrue, handleStarredFalse, handleDelete }) => {
+    const [finishedTasks, setFinishedTasks] = useState([])
     const rootUrl = process.env.REACT_APP_API_BASE_URL;
+    const { user } = useAuthContext()
+
     const fetchFinishedTasks = async () => {
-        const response = await fetch(`${rootUrl}/api/tasks/finished`);
+        const response = await fetch(`${rootUrl}/api/tasks/finished`, {
+            headers: { 'Authorization': `Bearer ${user.token}` }
+        });
         const json = await response.json();
 
         if (response.ok) {
-            setTasks(json);
+            setFinishedTasks(json);
         }
     }
 
     const handleDeleteAllFinished = async () => {
+        if (!user) {
+            return
+        }
+
         const response = await fetch(`${rootUrl}/api/tasks`, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${user.token}`
+            }
         })
         const json = await response.json()
 
@@ -33,8 +46,10 @@ const Finished = ({ tasks, setTasks, handleCheckTrue, handleCheckFalse, handleSt
     }
 
     useEffect(() => {
-        fetchFinishedTasks()
-    }, [fetchFinishedTasks])
+        if (user) {
+            fetchFinishedTasks()
+        }
+    }, [fetchFinishedTasks, user])
 
     return (
         <InfiniteScroll
@@ -51,13 +66,14 @@ const Finished = ({ tasks, setTasks, handleCheckTrue, handleCheckFalse, handleSt
                     <button className="all-delete-button" onClick={() => handleDeleteAllFinished()}>
                         <p>delete all</p>
                     </button>
-                    <div className="iconandtext">
+                    {/* <div className="iconandtext">
                         <BiSortAlt2 className="title-icon2" />
                         <h4 className="sort-title">Sort</h4>
-                    </div>
+                    </div> */}
+                    <div className="count"></div>
                 </div>
                 {/* <AddStarredTask fetchTasks={fetchTasks} /> */}
-                {tasks && tasks.map((task) => {
+                {finishedTasks && finishedTasks.map((task) => {
                     return <Task task={task} key={task._id} handleCheckTrue={handleCheckTrue} handleCheckFalse={handleCheckFalse} handleStarredTrue={handleStarredTrue} handleStarredFalse={handleStarredFalse} handleDelete={handleDelete} />
                 })}
             </div>
